pyproject.toml
[tool.poetry]
name = "Company test server"
version = "0.1.0"
description = "Test server for all project"
authors = ["ms <ff9147782@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pytoml = "^0.1.21"
redis = "^5.0.1"
aioredis = "2.0.1"
types-redis = "^4.6.0.11"

django = "^4.2.7"
django-cors-headers = "^4.3.1"
django-jinja = "^2.11.0"
django-phonenumber-field = "^7.2.0"
django-redis = "^5.4.0"
djangorestframework = "^3.14.0"

celery = "^5.3.6"
celery-progress = "^0.3"
celery-types = "^0.20.0"
django-celery-results = "^2.5.1"

loguru = "0.7.0"
phonenumberslite = "^8.13.25"
pillow = "^10.1.0"

#aiohttp = "3.8.6"
#aiohttp-proxy = "0.1.2"
psycopg2-binary = "^2.9.9"

art = "^6.1"
colorama = "^0.4.6"
types-colorama = "^0.4.15.12"

#aiofiles = "^23.2.1"
#matplotlib = "^3.8.2"
#httpx = {extras = ["socks"], version = "^0.27.0"}


[tool.poetry.group.development.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pylint = "^3.0.2"
django-debug-toolbar = "^4.2.0"
mypy = "^1.6.1"
pytz = "^2024.1"
types-pytz = "^2024.1.0.20240203"
coverage = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
pretty = true